+++ src/pages/signup/index.tsx
@@ -2,36 +2,35 @@ import { useRouter } from "next/router";
 import { postSignUp } from "@/api/auth";
 import { AxiosError } from "axios";
 import useNotify from "@/hooks/useNotify";
-import SignUpForm from "@/containers/SignUpForm";
-import { InputValues } from "@/containers/SignUpForm";
+import SignUpFormContainer from "@/containers/SignUpFormContainer";
+import { InputValues } from "@/containers/SignUpFormContainer";
 
 const SignUp = () => {
+  const router = useRouter();
+  const notify = useNotify();
 
-const router = useRouter();
-const notify = useNotify();
-
-const onSubmit = async (data: InputValues) => {
-  try {
-    await postSignUp({
-      email: data.email,
-      name: data.name,
-      password: data.password,
-      passwordConfirmation: data.passwordConfirmation,
-    });
-    notify("가입이 완료되었습니다.", "success");
-    router.push("/login");
-  } catch (err) {
-    if (err instanceof AxiosError) {
-      const msg =
-        err.response?.status === 400
-          ? err.response.data.message
-          : "이미 사용 중인 이름입니다.";
-      notify(msg, "error");
+  const onSubmit = async (data: InputValues) => {
+    try {
+      await postSignUp({
+        email: data.email,
+        name: data.name,
+        password: data.password,
+        passwordConfirmation: data.passwordConfirmation,
+      });
+      notify("가입이 완료되었습니다.", "success");
+      router.push("/login");
+    } catch (err) {
+      if (err instanceof AxiosError) {
+        const msg =
+          err.response?.status === 400
+            ? err.response.data.message
+            : "이미 사용 중인 이름입니다.";
+        notify(msg, "error");
+      }
     }
-  }
-};
+  };
 
-  return <SignUpForm onSubmit={onSubmit} />;
+  return <SignUpFormContainer onSubmit={onSubmit} />;
 };
 
 export default SignUp;
